Fix double-toggle when invoked on React editor
For the time being, we need to wrap React editor components in an extra
div that also has the editor class. This will go away when we're done
transitioning, but for now, this is a workaround that prevents the event
from being invoked when it matches on both `.editor` elements.